Fixed NPE, small refactors around namings
